+++ actix-tls/CHANGES.md
@@ -3,6 +3,14 @@
 ## Unreleased - 2021-xx-xx
 
 
+## 3.0.0-beta.4 - 2021-02-24
+* Rename `accept::openssl::{SslStream => TlsStream}`.
+* Add `connect::Connect::set_local_addr` to attach local `IpAddr`. [#282]
+* `connector::TcpConnector` service will try to bind to local_addr of `IpAddr` when given. [#282]
+
+[#282]: https://github.com/actix/actix-net/pull/282
+
+
 ## 3.0.0-beta.3 - 2021-02-06
 * Remove `trust-dns-proto` and `trust-dns-resolver`. [#248]
 * Use `std::net::ToSocketAddrs` as simple and basic default resolver. [#248]